Update count 3 length Strings.py
Given three integers a, b and c which denotes the frequencies of three different characters 'A', 'B', and 'C' respectively, and can be used to form strings of length 3. The task is to count the total number of possible combinations of A, B and C such that it forms a string having at least 2 different characters.

Example:

Input: a = 2, b = 3, c = 3
Output: 2
Explanation: Possible strings which satisfies the given conditions are: {"ABC", "ABC"}

Input: a = 5, b = 4, c = 3
Output: 4

 

Approach: Total number of strings of length 3, that can be formed with the given frequencies is (a+b+c)/3, assuming to select any character for any string. But as only strings with 2 different characters are required, so it's essential to check that if that's possible or not. To check that:

Assume that, all (a+b+c)/3 strings are formed up to two places only with any and all does have a remaining space left to be filled.Now, till this point are strings are valid because:If the string has two different characters, then it's valid.If the string has two same characters, then it can be made valid by inserting a different character.So, the total number of different characters we need is, let's say count where count = (a+b+c)/3, assuming that one character is required for each string.So if the sum of the two smallest frequencies exceeds count, then (a+b+c)/3 number of strings can be formed. Otherwise, it would be the sum of two smallest frequencies.